Fix bug where turn sink was canceled after game end
Change server-side Flux game events turn sink to a hot source so that it is not canceled when one or more publishers are canceled by the WebSocket handler. Previously, closing the WebSocket on the frontend would cause cancellation of the turn sink. This prevented future clients from connecting to the game events WebSocket.