Add script to deploy local dev DynamoDB schema
Service configuration looks for DynamoDB table names from environment variables. This supports full AWS and local dev deployment. A python script in the deployment folder will add tables to the local DynamoDB docker instance based on the CloudFormation template, and generates a file that can be sourced to define the required environment variables.